refactor: restructure interface-implement #27
