+++ src/main/java/com/kospot/kospot/domain/coordinate/adaptor/CoordinateAdaptor.java
@@ -2,18 +2,66 @@ package com.kospot.kospot.domain.coordinate.adaptor;
 
 import com.kospot.kospot.domain.coordinate.entity.Coordinate;
 import com.kospot.kospot.domain.coordinate.entity.sido.Sido;
+import com.kospot.kospot.domain.coordinate.repository.CoordinateNationwideRepository;
+import com.kospot.kospot.domain.coordinate.service.DynamicCoordinateRepositoryFactory;
+import com.kospot.kospot.exception.object.domain.CoordinateHandler;
+import com.kospot.kospot.exception.payload.code.ErrorStatus;
 
-public interface CoordinateAdaptor {
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-    Coordinate queryById(Sido sido, Long id);
+import java.util.Optional;
 
-    boolean queryExistsById(Sido sido, Long id);
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class CoordinateAdaptor implements CoordinateAdaptor{
 
-    Long queryMaxIdBySido(Sido sido);
+    private final DynamicCoordinateRepositoryFactory factory;
+    private final CoordinateNationwideRepository nationwideRepository;
 
-    Coordinate queryNationwideById(Long id);
+    @Override
+    public Coordinate queryById(Sido sido, Long id) {
+        return factory.getRepository(sido).findById(id)
+                .map(Coordinate.class::cast)
+                .orElseThrow(
+                () -> new CoordinateHandler(ErrorStatus.COORDINATE_NOT_FOUND)
+        );
+    }
 
-    Long queryNationwideMaxId();
+    @Override
+    public boolean queryExistsById(Sido sido, Long id) {
+        return factory.getRepository(sido).existsById(id);
+    }
 
-    boolean queryNationwideExistsById(Long id);
+    @Override
+    public Long queryMaxIdBySido(Sido sido){
+        Long maxId = factory.getRepository(sido).findMaxId();
+        return Optional.ofNullable(maxId).orElse(0L);
+    }
+
+    /**
+     *  Nationwide
+     */
+
+    @Override
+    public Coordinate queryNationwideById(Long id) {
+        return nationwideRepository.findById(id)
+                .map(Coordinate.class::cast)
+                .orElseThrow(
+                        () -> new CoordinateHandler(ErrorStatus.COORDINATE_NOT_FOUND)
+                );
+    }
+
+    @Override
+    public boolean queryNationwideExistsById(Long id) {
+        return nationwideRepository.existsById(id);
+    }
+
+    @Override
+    public Long queryNationwideMaxId() {
+        Long maxId = nationwideRepository.findMaxId();
+        return Optional.ofNullable(maxId).orElse(0L);
+    }
 }
