+++ src/main/java/com/kospot/kospot/domain/coordinateIdCache/service/CoordinateIdCacheService.java
@@ -1,8 +1,38 @@
 package com.kospot.kospot.domain.coordinateIdCache.service;
 
+import com.kospot.kospot.domain.coordinate.entity.sido.Sido;
+import com.kospot.kospot.domain.coordinateIdCache.entity.CoordinateIdCache;
+import com.kospot.kospot.domain.coordinateIdCache.repository.CoordinateIdCacheRepository;
 
-public interface CoordinateIdCacheService {
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-    void saveAllMaxId();
+@Service
+@RequiredArgsConstructor
+public class CoordinateIdCacheService implements CoordinateIdCacheService {
+
+    private final CoordinateAdaptor coordinateAdaptor;
+    private final CoordinateIdCacheRepository repository;
+
+    @Override
+    @Transactional
+    public void saveAllMaxId() {
+        for (Sido sido : Sido.values()) {
+            saveMaxIdBySido(sido);
+        }
+    }
+
+    public void saveMaxIdBySido(Sido sido) {
+        Long maxId = coordinateAdaptor.queryMaxIdBySido(sido);
+
+        repository.findById(sido).ifPresentOrElse(
+                coordinateIdCache -> coordinateIdCache.updateMaxId(maxId),
+                () -> repository.save(CoordinateIdCache.builder()
+                        .sido(sido)
+                        .maxId(maxId)
+                        .build())
+        );
+    }
 
 }
