+++ src/main/java/com/kospot/kospot/domain/game/controller/RoadViewGameController.java
@@ -3,8 +3,8 @@ package com.kospot.kospot.domain.game.controller;
 import com.kospot.kospot.domain.game.dto.request.EndGameRequest;
 import com.kospot.kospot.domain.game.dto.response.EndGameResponse;
 import com.kospot.kospot.domain.game.dto.response.StartGameResponse;
-import com.kospot.kospot.domain.game.service.RoadViewGameService;
 import com.kospot.kospot.domain.game.service.AESService;
+import com.kospot.kospot.domain.game.service.RoadViewGameService;
 import com.kospot.kospot.domain.game.util.ScoreCalculator;
 import com.kospot.kospot.domain.member.entity.Member;
 import com.kospot.kospot.exception.payload.dto.ApiResponseDto;
