--- src/main/java/com/kospot/kospot/domain/member/adaptor/MemberAdaptorImpl.java
@@ -1,25 +1,0 @@
-package com.kospot.kospot.domain.member.adaptor;
-
-import com.kospot.kospot.domain.member.entity.Member;
-import com.kospot.kospot.domain.member.repository.MemberRepository;
-import com.kospot.kospot.exception.object.domain.MemberHandler;
-import com.kospot.kospot.exception.payload.code.ErrorStatus;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class MemberAdaptorImpl implements MemberAdaptor{
-
-    private final MemberRepository repository;
-
-    @Override
-    public Member queryById(Long memberId) {
-        return repository.findById(memberId).orElseThrow(
-                () -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND)
-        );
-    }
-
-}
