+++ src/main/java/com/kospot/kospot/domain/point/adaptor/PointHistoryAdaptor.java
@@ -1,10 +1,23 @@
 package com.kospot.kospot.domain.point.adaptor;
 
+
 import com.kospot.kospot.domain.point.entity.PointHistory;
+import com.kospot.kospot.domain.point.repository.PointHistoryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface PointHistoryAdaptor {
-    List<PointHistory> queryAllHistoryByMemberId(Long memberId);
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class PointHistoryAdaptor {
+
+    private final PointHistoryRepository repository;
+
+    public List<PointHistory> queryAllHistoryByMemberId(Long memberId) {
+        return repository.findAllByMemberId(memberId);
+    }
 
 }
