+++ src/main/java/com/kospot/kospot/domain/point/service/PointHistoryService.java
@@ -1,15 +1,37 @@
 package com.kospot.kospot.domain.point.service;
 
 import com.kospot.kospot.domain.member.entity.Member;
+import com.kospot.kospot.domain.point.adaptor.PointHistoryAdaptor;
 import com.kospot.kospot.domain.point.dto.response.PointHistoryResponse;
+import com.kospot.kospot.domain.point.entity.PointHistory;
 import com.kospot.kospot.domain.point.entity.PointHistoryType;
+import com.kospot.kospot.domain.point.repository.PointHistoryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
-public interface PointHistoryService {
+@Service
+@RequiredArgsConstructor
+public class PointHistoryService {
 
-    List<PointHistoryResponse> findAllHistoryByMemberId(Long memberId);
+    private final PointHistoryRepository repository;
+    private final PointHistoryAdaptor adaptor;
 
-    void savePointHistory(Member member, int amount, PointHistoryType pointHistoryType);
+    public List<PointHistoryResponse> findAllHistoryByMemberId(Long memberId) {
+        List<PointHistory> histories = adaptor.queryAllHistoryByMemberId(memberId);
+        return histories.stream()
+                .map(PointHistoryResponse::from)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public void savePointHistory(Member member, int amount, PointHistoryType pointHistoryType){
+        repository.save(
+                PointHistory.create(member, amount, pointHistoryType)
+        );
+    }
 
 }
